feat: create api gateway for morphology; ecr repository; deployment workflow
