+++ .github/workflows/deploy_core.yml
@@ -7,7 +7,7 @@ on:
     paths:
       - src/**
       - charts/grammr/**
-      - charts/values/**
+      - config.yaml
       - pom.xml
       - Dockerfile
       - .github/workflows/deploy_core.yml
