+++ .github/workflows/deploy_morphology_serverless.yaml
@@ -1,0 +1,69 @@
+name: Build Docker Images
+
+on:
+  pull_request:
+    branches:
+      - main
+  push:
+    branches: [main]
+    paths:
+      - 'config.yaml'
+      - 'Dockerfile'
+      - '.github/workflows/build-images.yml'
+  workflow_dispatch:
+
+jobs:
+  prepare:
+    runs-on: ubuntu-latest
+    outputs:
+      matrix: ${{ steps.set-matrix.outputs.matrix }}
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v3
+
+      - name: Parse YAML and create build matrix
+        id: set-matrix
+        run: |
+          # Install yq for YAML parsing
+          sudo wget -qO /usr/local/bin/yq https://github.com/mikefarah/yq/releases/latest/download/yq_linux_amd64
+          sudo chmod +x /usr/local/bin/yq
+          
+          MATRIX=$(yq eval -o=json '.languages' config.yaml)
+          echo "matrix=${MATRIX}" >> $GITHUB_OUTPUT
+          echo "Generated matrix: ${MATRIX}"
+
+  build:
+    needs: prepare
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        language: ${{ fromJson(needs.prepare.outputs.matrix) }}
+      # Allow other jobs to continue even if one fails
+      fail-fast: false
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v3
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v2
+
+      - name: Configure AWS Credentials
+        uses: aws-actions/configure-aws-credentials@v4
+        with:
+          role-to-assume: arn:aws:iam::246770851643:role/github-actions
+          aws-region: eu-central-1
+
+      - name: Login to Amazon ECR
+        id: login-ecr
+        uses: aws-actions/amazon-ecr-login@v1
+
+      - name: Build and push Docker image
+        uses: docker/build-push-action@v4
+        with:
+          context: .
+          push: true
+          tags: 246770851643.dkr.ecr.eu-central-1.amazonaws.com/morphology:latest-${{ matrix.language.code }}
+          build-args: |
+            MODEL=${{ matrix.language.model }}
