+++ sidecars/morphology/docker/init.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 echo "Installing model ${SPACY_MODEL}"
-python3 -m spacy download "${SPACY_MODEL}"
+
 
 echo "Starting FastAPI server"
 python3 -m uvicorn analysis.main:app --host=0.0.0.0 --port=8000
