+++ sidecars/morphology/morphology/main.py
@@ -2,8 +2,8 @@ import logging
 from fastapi import FastAPI
 from fastapi.responses import JSONResponse
 
-from analysis.domain.analysis_request import AnalysisRequest
-from analysis.service import analysis_service
+from morphology.domain.analysis_request import AnalysisRequest
+from morphology.service import analysis_service
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(
@@ -16,7 +16,7 @@ app = FastAPI()
 async def analyze(request: AnalysisRequest) -> JSONResponse:
     logger.info(f"Received request: {request}")
     analysis = analysis_service.perform_analysis(request)
-    logger.info(f"Returning analysis: {analysis.model_dump()}")
+    logger.info(f"Returning morphology: {analysis.model_dump()}")
     return JSONResponse(
         content=analysis.model_dump(), headers={"Content-Type": "application/json"}
     )
