+++ sidecars/morphology/pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "analysis"
+name = "morphology"
 version = "0.2.0"
 description = ""
 authors = ["Tobias Waslowski <tobiaswaslowski@gmail.com>"]
@@ -7,6 +7,7 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.12"
+# todo: i know this was required for SOME compatibility reasons. figure out the specifics.
 spacy = "3.7.0"
 uvicorn = "^0.32.0"
 fastapi = "^0.115.4"
