+++ terraform/ecr.tf
@@ -1,0 +1,32 @@
+// module "syntactical_analysis_repository" {
+//   source = "terraform-aws-modules/ecr/aws"
+//
+//   repository_name                 = "syntactical_analysis-lambda-${var.environment}"
+//   repository_image_tag_mutability = "MUTABLE"
+//
+//   repository_read_write_access_arns = [
+//     data.aws_caller_identity.current.arn,
+//     module.morphology.lambda_role_arn
+//   ]
+//
+//   repository_lifecycle_policy = local.repository_lifecycle_policy
+// }
+//
+// locals {
+//   repository_lifecycle_policy = jsonencode({
+//     rules = [
+//       {
+//         rulePriority = 1,
+//         description  = "Keep 3 images",
+//         selection = {
+//           tagStatus   = "any",
+//           countType   = "imageCountMoreThan",
+//           countNumber = 2
+//         },
+//         action = {
+//           type = "expire"
+//         }
+//       }
+//     ]
+//   })
+// }
