+++ terraform/local.tf
@@ -1,5 +1,5 @@
 locals {
-  config = yamldecode(file("${path.module}/../charts/values/grammr.values.yaml"))
+  config = yamldecode(file("${path.module}/../config.yaml"))
 
   languages_map = {
     for lang in local.config.languages : lang.code => lang
