+++ terraform/modules/lambda/api_gateway.tf
@@ -1,0 +1,21 @@
+resource "aws_api_gateway_resource" "resource" {
+  rest_api_id = var.api_gateway_id
+  parent_id   = var.root_resource_id
+  path_part   = var.name
+}
+
+resource "aws_api_gateway_method" "post" {
+  rest_api_id   = var.api_gateway_id
+  resource_id   = aws_api_gateway_resource.resource.id
+  http_method   = "POST"
+  authorization = "NONE"
+}
+
+resource "aws_api_gateway_integration" "lambda_integration" {
+  rest_api_id             = var.api_gateway_id
+  resource_id             = aws_api_gateway_resource.resource.id
+  http_method             = aws_api_gateway_method.post.http_method
+  integration_http_method = "POST"
+  type                    = "AWS_PROXY"
+  uri                     = module.lambda.lambda_function_invoke_arn
+}
